+++ jumpscale/packages/marketplace/chats/domain.py
@@ -3,7 +3,7 @@ import requests
 from jumpscale.god import j
 from jumpscale.sals.chatflows.chatflows import GedisChatBot, StopChatFlow, chatflow_step
 from jumpscale.sals.reservation_chatflow.models import SolutionType
-from jumpscale.sals.deployer import deployer, MarketPlaceChatflow
+from jumpscale.sals.marketplace import deployer, MarketPlaceChatflow
 
 
 class DomainDelegation(MarketPlaceChatflow):
